Implement the display driver for the v1.
I rewrote it to be based on a hardware timer instead of async/await,
because the previous implementation was too slow for the v1's slower processor.

There can be some slight flickering of dim pixels with this right now on the v1,
caused by the frame being set, which will disable the interrupt temporarily and make it run later than it should, making any currently-rendering pixels brighter than they should be.
This can be worked around by avoiding disabling the interrupt there, e.g. by using a mutex but just giving up if it's locked, but this is good enough for the moment.